cleaned up info
