+++ src/lib.rs
@@ -7,7 +7,6 @@ mod http;
 mod library;
 mod lockfile;
 mod package;
-mod plan_summary;
 mod project_info;
 mod r_cmd;
 mod renv;
@@ -30,7 +29,6 @@ pub use http::{Http, HttpDownload};
 pub use library::Library;
 pub use lockfile::Lockfile;
 pub use package::{is_binary_package, Version, VersionRequirement};
-pub use plan_summary::DependencyInfo;
 pub use project_info::ProjectInfo;
 pub use r_cmd::{find_r_version_command, RCmd, RCommandLine};
 pub use renv::RenvLock;
