+++ src/main.rs
@@ -6,7 +6,7 @@ use fs_err as fs;
 use rv::cli::utils::timeit;
 use rv::cli::{find_r_repositories, init, migrate_renv, CliContext};
 use rv::{
-    activate, deactivate, CacheInfo, DependencyInfo, Git, Http, Lockfile, ProjectInfo, RCmd,
+    activate, deactivate, CacheInfo, Git, Http, Lockfile, ProjectInfo, RCmd,
     RCommandLine, ResolvedDependency, Resolver, SyncHandler,
 };
 
@@ -232,24 +232,15 @@ fn try_main() -> Result<()> {
             let mut context = CliContext::new(&cli.config_file)?;
             context.load_databases()?;
             let resolved = resolve_dependencies(&context);
-            let info = DependencyInfo::new(
-                &context.library,
-                &resolved,
-                &context.config.repositories(),
-                &context.databases,
-                &context.r_version,
-                &context.cache,
-                context.lockfile.as_ref(),
-            );
-            println!("{info}");
-            // if json {
-            //     println!(
-            //         "{}",
-            //         serde_json::to_string_pretty(&info).expect("valid json")
-            //     );
-            // } else {
-            //     println!("{info}");
-            // }
+            let info = ProjectInfo::new(&context.library, &resolved, &context.config.repositories(), &context.databases, &context.r_version, &context.cache, context.lockfile.as_ref());
+            if json {
+                println!(
+                    "{}",
+                    serde_json::to_string_pretty(&info).expect("valid json")
+                );
+            } else {
+                println!("{info}");
+            }
         }
         Command::Activate => {
             let dir = std::env::current_dir()?;
