+++ src/project_info.rs
@@ -4,272 +4,35 @@ use std::{
     path::Path,
 };
 
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 
 use crate::{
-    cache::InstallationStatus, lockfile::Source, package::PackageType, DiskCache, Library,
-    Lockfile, Repository, RepositoryDatabase, ResolvedDependency, SystemInfo, Version,
-    VersionRequirement,
+    cache::InstallationStatus, lockfile::Source, package::{Operator, PackageType}, DiskCache, Library, Lockfile, Repository, RepositoryDatabase, ResolvedDependency, SystemInfo, Version, VersionRequirement
 };
-
-#[derive(Debug, Clone, Serialize, Default)]
-struct SourceCounts {
-    total: usize,
-    installed: usize,
-    in_cache: usize,
-    to_download: usize,
-}
-
-impl SourceCounts {
-    fn update(&mut self, dep_status: DependencyStatus) {
-        self.total += 1; // all updates increment total
-        match dep_status {
-            DependencyStatus::InCache => self.in_cache += 1,
-            DependencyStatus::NotInCache => self.to_download += 1,
-            DependencyStatus::Installed => self.installed += 1,
-        }
-    }
-}
-
-// used to communicate which SourceCounts to update
-// Cache variants are only used for packages that are not installed
-#[derive(Debug, Clone)]
-enum DependencyStatus {
-    InCache,
-    NotInCache,
-    Installed,
-}
-
-fn update_counts<'a>(
-    counts: &mut HashMap<&'a str, (SourceCounts, SourceCounts)>,
-    source_id: &'a str,
-    dep_status: DependencyStatus,
-    is_binary: bool,
-) {
-    let (source_counts, binary_counts) = counts
-        .entry(source_id)
-        .or_insert((SourceCounts::default(), SourceCounts::default()));
-    if is_binary {
-        binary_counts.update(dep_status);
-    } else {
-        source_counts.update(dep_status);
-    }
-}
-
-#[derive(Debug, Clone, Serialize)]
-struct DependencyInfo<'a> {
-    path: &'a Path,
-    total: usize,
-    counts: HashMap<&'a str, (SourceCounts, SourceCounts)>,
-    to_remove: usize,
-    non_locked: Option<usize>, // None iff lockfile is not present
-}
-
-impl<'a> DependencyInfo<'a> {
-    pub fn new(
-        library_dir: &'a Path,
-        lockfile: &'a Option<Lockfile>,
-        cache: &'a DiskCache,
-        library: &'a Library,
-        resolved_dependencies: &'a [ResolvedDependency],
-        repo_dbs: &[(RepositoryDatabase, bool)],
-        repos: &'a [Repository],
-        r_version: &Version,
-    ) -> Self {
-        let mut counts = HashMap::new();
-        let mut non_locked = 0;
-        let mut to_remove = 0;
-        let mut lib_clone = library
-            .packages
-            .keys()
-            .map(|s| s.to_string())
-            .collect::<HashSet<String>>();
-
-        let mut source_id_hash = HashMap::new();
-        for r in resolved_dependencies {
-            let source_id = source_id_hash
-                .entry(&r.source)
-                .or_insert(get_source_id(r, &repos));
-            let is_binary = is_package_from_binary(r, repo_dbs, r_version);
-
-            if library.contains_package(&r.name, Some(&r.version)) {
-                lib_clone.remove(r.name.as_ref());
-                update_counts(
-                    &mut counts,
-                    &source_id,
-                    DependencyStatus::Installed,
-                    is_binary,
-                );
-            } else {
-                match cache.get_installation_status(&r.name, &r.version.original, &r.source) {
-                    // If the package is in the cache as a binary, we want to record it as a binary, no matter if the resolved dependency is from source or binary
-                    // This is because we want to use this as a way to convey to the user the performance costs to perform the installation
-                    InstallationStatus::Both | InstallationStatus::Binary => {
-                        update_counts(&mut counts, &source_id, DependencyStatus::InCache, true)
-                    }
-                    // We only want to say a source package is in the cache if the resolved dependency is from source.
-                    // If the dep is supposed to be from binary, we prefer to download a binary rather than build our own
-                    InstallationStatus::Source if !is_binary => {
-                        update_counts(&mut counts, &source_id, DependencyStatus::InCache, false)
-                    }
-                    // If the package is Absent or Source, but we want a binary, we record it as not in the cache
-                    _ => update_counts(
-                        &mut counts,
-                        &source_id,
-                        DependencyStatus::NotInCache,
-                        is_binary,
-                    ),
-                }
-            }
-        }
-
-        for name in lib_clone {
-            if let Some(lock) = lockfile {
-                if lock.get_package(&name, None).is_some() {
-                    to_remove += 1;
-                } else {
-                    non_locked += 1;
-                }
-            } else {
-                to_remove += 1;
-            }
-        }
-
-        let non_locked = lockfile.as_ref().map(|_| non_locked);
-
-        Self {
-            path: library_dir,
-            total: resolved_dependencies.len(),
-            counts,
-            to_remove,
-            non_locked,
-        }
-    }
-}
-
-impl fmt::Display for DependencyInfo<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "Installed: {}/{}\n{}{}\n",
-            self.counts
-                .values()
-                .map(|(s, b)| s.installed + b.installed)
-                .sum::<usize>(),
-            self.total,
-            format!("To remove: {}\n", self.to_remove).repeat((self.to_remove != 0) as usize),
-            match self.non_locked {
-                Some(n) if n != 0 => format!("Packages not in lock file: {n}\n"),
-                _ => String::new(),
-            }
-        )?;
-
-        write!(f, "Package Sources:\n")?;
-        let mut install_needed = false; // If all packages are installed, no need to show installation summary
-        let mut counts_vec: Vec<_> = self.counts.iter().collect();
-        counts_vec.sort_by(|a, b| a.0.cmp(b.0));
-        for (s, (source_counts, binary_counts)) in &counts_vec {
-            if source_counts.total != source_counts.installed
-                || binary_counts.total != binary_counts.installed
-            {
-                install_needed = true;
-            }
-            write!(
-                f,
-                "  {}: {}{}{}\n",
-                s,
-                format!(
-                    "{}/{} binary packages",
-                    binary_counts.installed, binary_counts.total
-                )
-                .repeat((binary_counts.total != 0) as usize),
-                ", ".repeat((binary_counts.total != 0 && source_counts.total != 0) as usize),
-                format!(
-                    "{}/{} source packages",
-                    source_counts.installed, source_counts.total
-                )
-                .repeat((source_counts.total != 0) as usize)
-            )?;
-        }
-
-        if !install_needed {
-            return Ok(());
-        }
-        write!(f, "\nInstallation Summary:\n")?;
-        for (s, (source_counts, binary_counts)) in &counts_vec {
-            let diff = binary_counts.total - binary_counts.installed + source_counts.total
-                - source_counts.installed;
-            let in_cache = binary_counts.in_cache + source_counts.in_cache;
-            let to_download = binary_counts.to_download + source_counts.to_download;
-            if diff == 0 {
-                continue;
-            }
-            write!(
-                f,
-                "  {}: {}{}{}\n",
-                s,
-                format!(
-                    "{}/{} in cache{}",
-                    in_cache,
-                    diff,
-                    format!(" ({} require compilation)", source_counts.in_cache)
-                        .repeat((source_counts.in_cache != 0) as usize)
-                )
-                .repeat((in_cache != 0) as usize),
-                ", ".repeat((in_cache != 0 && to_download != 0) as usize),
-                format!(
-                    "{}/{} to download{}",
-                    to_download,
-                    diff,
-                    format!(" ({} require compilation)", source_counts.to_download)
-                        .repeat((source_counts.to_download != 0) as usize)
-                )
-                .repeat((to_download != 0) as usize),
-            )?;
-        }
-        Ok(())
-    }
-}
 
 #[derive(Debug, Clone, Serialize)]
 pub struct ProjectInfo<'a> {
-    system_info: &'a SystemInfo,
     r_version: &'a Version,
-    dep_info: DependencyInfo<'a>,
-    remote_info: RemoteInfo,
+    system_info: &'a SystemInfo,
+    dependency_info: DependencyInfo<'a>,
+    remote_info: RemoteInfo<'a>,
 }
 
 impl<'a> ProjectInfo<'a> {
     pub fn new(
-        lockfile: &'a Option<Lockfile>,
-        cache: &'a DiskCache,
-        r_version: &'a Version,
         library: &'a Library,
-        repository_databases: &'a [(RepositoryDatabase, bool)],
-        resolved_dependencies: &'a [ResolvedDependency],
+        resolved_deps: &'a [ResolvedDependency],
         repositories: &'a [Repository],
+        repo_dbs: &'a [(RepositoryDatabase, bool)],
+        r_version: &'a Version,
+        cache: &'a DiskCache,
+        lockfile: Option<&'a Lockfile>,
     ) -> Self {
-        let dep_info = DependencyInfo::new(
-            library.path(),
-            lockfile,
-            cache,
-            library,
-            resolved_dependencies,
-            &repository_databases,
-            repositories,
-            r_version,
-        );
-        let remote_info: RemoteInfo = RemoteInfo::new(
-            &r_version.major_minor(),
-            repository_databases,
-            &repositories,
-        );
-
         Self {
+            r_version,
             system_info: &cache.system_info,
-            r_version: r_version,
-            dep_info,
-            remote_info,
+            dependency_info: DependencyInfo::new(library, resolved_deps, repositories, repo_dbs, r_version, cache, lockfile),
+            remote_info: RemoteInfo::new(repositories, repo_dbs, &r_version.major_minor()),
         }
     }
 }
@@ -277,119 +40,365 @@ impl<'a> ProjectInfo<'a> {
 impl fmt::Display for ProjectInfo<'_> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(
-            f,
-            "== System Info ==\nOS: {}{}\nR Version: {}\n",
+            f, 
+            "OS: {}{}\nR Version: {}\n\n",
             self.system_info.os_family(),
-            self.system_info
-                .arch()
-                .map(|a| format!(" ({a})"))
-                .unwrap_or_default(),
+            if let Some(arch) = self.system_info.arch() {
+                format!(" ({arch})")
+            } else {
+                String::new()
+            },
             self.r_version,
         )?;
 
+        write!(f, "== Dependencies == \n{}\n", self.dependency_info)?;
+        write!(f, "== Remote == \n{}", self.remote_info)?;
+        Ok(())
+    }
+}
+
+// TODO: Expand with git + url information
+#[derive(Debug, Clone, Serialize)]
+struct RemoteInfo<'a> {
+    repositories: HashMap<String, (&'a str, usize, usize)>
+}
+
+impl<'a> RemoteInfo<'a> {
+    fn new(repos: &'a [Repository], repo_dbs: &'a [(RepositoryDatabase, bool)], r_version: &[u32; 2]) -> Self {
+        let mut repositories = HashMap::new();
+        for (repo_db, _) in repo_dbs {
+            let bin_count = repo_db.get_binary_count(r_version);
+            let src_count = repo_db.get_source_count();
+            let id = get_repository_alias(&repo_db.url, repos);
+            repositories.insert(id, (repo_db.url.as_str(), bin_count, src_count));
+        }
+
+        Self {
+            repositories
+        }
+    }
+}
+
+impl fmt::Display for RemoteInfo<'_> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        for (alias, (url, bin_count, src_count)) in &self.repositories {
+            write!(f, "{alias} ({url}): {bin_count} binary packages, {src_count} source packages\n")?;
+        }
+        Ok(())
+    }
+}
+
+#[derive(Debug, Clone, Serialize)]
+struct DependencyInfo<'a> {
+    lib_path: &'a Path,
+    dependencies: HashMap<String, Vec<DependencySummary<'a>>>,
+    to_remove: HashSet<String>,
+    non_locked: HashSet<String>,
+}
+
+impl<'a> DependencyInfo<'a> {
+    fn new(
+        library: &'a Library,
+        resolved_deps: &'a [ResolvedDependency],
+        repositories: &'a [Repository],
+        repo_dbs: &[(RepositoryDatabase, bool)],
+        r_version: &Version,
+        cache: &'a DiskCache,
+        lockfile: Option<&'a Lockfile>,
+    ) -> Self {
+        let mut non_locked = HashSet::new();
+        let mut to_remove = HashSet::new();
+        let mut dependencies: HashMap<String, Vec<DependencySummary>> = HashMap::new();
+
+        // we keep a list of packages within the lib, removing each package as each dependency is processed
+        // any libs left in the list either need to be removed or are not locked
+        let mut lib_pkgs = library
+            .packages
+            .keys()
+            .map(|s| s.to_string())
+            .collect::<HashSet<_>>();
+
+        // we keep track of the dependencies organized by their source identifier
+        for r in resolved_deps {
+            lib_pkgs.remove(r.name.as_ref());
+            let mut dep_sum = DependencySummary::new(r, library, repo_dbs, r_version, cache);
+            // if the package was found in the library, but not in the lockfile, we consider it not locked and is missing
+            if !is_in_lock(r.name.as_ref(), lockfile) && dep_sum.status == DependencyStatus::Installed{
+                dep_sum.status = DependencyStatus::Missing;
+                non_locked.insert(r.name.to_string());
+                continue;
+            }
+            let dep_id = get_dep_id(r, repositories);
+            dependencies.entry(dep_id).or_default().push(dep_sum);
+        }
+
+        // Any packages still left in lib_pkgs are superfluous and should be removed
+        // Additionally, packages that are not in the lockfile need to be reported and additionally removed
+        for pkg in &lib_pkgs {
+            if is_in_lock(pkg, lockfile) {
+                non_locked.insert(pkg.to_string());
+            }
+            to_remove.insert(pkg.to_string());
+        }
+
+        Self {
+            lib_path: library.path(),
+            dependencies,
+            to_remove,
+            non_locked,
+        }
+    }
+
+    fn num_deps_total(&self) -> usize {
+        self.dependencies.values().flatten().count()
+    }
+
+    fn num_deps_installed(&self) -> usize {
+        self.dependencies
+            .values()
+            .flatten()
+            .filter(|d| d.status == DependencyStatus::Installed)
+            .count()
+    }
+}
+
+impl fmt::Display for DependencyInfo<'_> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(
             f,
-            "\n== Dependencies (Installed/Total) ==\n{}",
-            self.dep_info
+            "Library: {}\nInstalled: {}/{}\n{}{}\n",
+            self.lib_path.to_string_lossy(),
+            self.num_deps_installed(),
+            self.num_deps_total(),
+            when_non_zero(
+                &format!("To remove: {}\n", self.to_remove.len()),
+                self.to_remove.len()
+            ),
+            when_non_zero(
+                &format!("Not in lock file: {}\n", self.non_locked.len()),
+                self.non_locked.len()
+            )
         )?;
 
-        write!(f, "\n== Repositories ==\n{}", self.remote_info)?;
+        let mut pkg_source = String::from("Package Sources: \n");
+        let mut install_summary = String::from("\nInstallation Summary: \n");
+        for (s, dep_vec) in &self.dependencies {
+            let counts = Counts::new(dep_vec);
+            pkg_source.push_str(&format!(
+                "  {}: {}{}{}\n",
+                s,
+                when_non_zero(
+                    &format!("{}/{} binary packages", counts.installed_bin, counts.total_bin),
+                    counts.total_bin
+                ),
+                when_non_zero(", ", (counts.total_bin != 0 && counts.total_src != 0) as usize),
+                when_non_zero(
+                    &format!("{}/{} source packages", counts.installed_src, counts.total_src),
+                    counts.total_src
+                ),
+            ));
+            if counts.to_install == 0 {
+                continue;
+            }
+            install_summary.push_str(&format!(
+                "  {}: {}{}{}\n",
+                s,
+                when_non_zero(
+                    &format!(
+                        "{}/{} in cache{}",
+                        counts.in_cache, 
+                        counts.to_install,
+                        when_non_zero(&format!(" ({} to compile)", counts.in_cache_to_compile), counts.in_cache_to_compile)
+                    ),
+                    counts.in_cache
+                ),
+                when_non_zero(", ", (counts.in_cache != 0 && counts.to_download != 0) as usize),
+                when_non_zero(
+                    &format!(
+                        "{}/{} to download{}",
+                        counts.to_download, 
+                        counts.to_install,
+                        when_non_zero(&format!(" ({} to compile)", counts.to_download_to_compile), counts.to_download_to_compile)
+                    ),
+                    counts.to_download
+                )
+            ));
+
+        }
+        write!(f, "{pkg_source}")?;
+        write!(f, "{install_summary}")?;
 
         Ok(())
     }
 }
 
-fn is_package_from_binary(
+#[derive(Debug, Clone, PartialEq, Serialize)]
+enum DependencyStatus {
+    Installed,
+    InCache,
+    ToCompile,
+    Missing,
+}
+
+#[derive(Debug, Clone, Serialize)]
+struct DependencySummary<'a> {
+    _name: &'a str, // for eventual ability to list pkg names
+    is_binary: bool,
+    status: DependencyStatus,
+}
+
+impl<'a> DependencySummary<'a> {
+    pub fn new(
+        resolved_dep: &'a ResolvedDependency,
+        library: &Library,
+        repo_dbs: &[(RepositoryDatabase, bool)],
+        r_version: &Version,
+        cache: &DiskCache,
+    ) -> Self {
+        let is_binary = is_binary_package(resolved_dep, repo_dbs, r_version);
+
+        if library.contains_package(&resolved_dep.name, Some(&resolved_dep.version)) {
+            return Self {
+                _name: &resolved_dep.name,
+                is_binary,
+                status: DependencyStatus::Installed,
+            };
+        };
+
+        let status = match cache.get_installation_status(
+            &resolved_dep.name,
+            &resolved_dep.version.original,
+            &resolved_dep.source,
+        ) {
+            // If the package has a binary in the cache, we can use it independent of if the package is binary or not
+            InstallationStatus::Both | InstallationStatus::Binary => DependencyStatus::InCache,
+            // If the dependency is not a binary and we have the source in the cache, we can compile it
+            InstallationStatus::Source if !is_binary => DependencyStatus::ToCompile,
+            // If the dependency is absent or only source when we want a binary, we report it as missing
+            _ => DependencyStatus::Missing,
+        };
+
+        Self {
+            _name: &resolved_dep.name,
+            is_binary,
+            status,
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy)]
+struct Counts {
+    total_bin: usize,
+    total_src: usize,
+    installed_bin: usize,
+    installed_src: usize,
+    to_install: usize,
+    in_cache: usize,
+    in_cache_to_compile: usize,
+    to_download: usize,
+    to_download_to_compile: usize,
+}
+
+impl Counts {
+    fn new(dep_vec: &[DependencySummary]) -> Self {
+        let mut counts = Counts {
+            total_bin: 0,
+            total_src: 0,
+            installed_bin: 0,
+            installed_src: 0,
+            to_install: 0,
+            in_cache: 0,
+            in_cache_to_compile: 0,
+            to_download: 0,
+            to_download_to_compile: 0,
+        };
+
+        for dep in dep_vec {
+            if dep.is_binary {
+                counts.total_bin += 1;
+                if let DependencyStatus::Installed = dep.status {
+                    counts.installed_bin += 1;
+                    continue;
+                }
+            } else {
+                counts.total_src += 1;
+                match dep.status {
+                    DependencyStatus::Installed => {
+                        counts.installed_src += 1;
+                        continue;
+                    }
+                    DependencyStatus::ToCompile => counts.in_cache_to_compile += 1,
+                    DependencyStatus::Missing => counts.to_download_to_compile += 1,
+                    _ => (),
+                }
+            }
+            counts.to_install += 1;
+            match dep.status {
+                DependencyStatus::InCache => counts.in_cache += 1,
+                DependencyStatus::Missing => counts.to_download += 1,
+                _ => (),
+            }
+        }
+        counts
+    }
+}
+
+fn when_non_zero(s: &str, arg_of_interest: usize) -> &str {
+    if arg_of_interest != 0 {
+        s
+    } else {
+        ""
+    }
+}
+
+// Determine if pkg is in the lockfile, if lockfile is None, we assume all packages are in the lockfile
+// This is because we are using if a package is not in a lockfile as a proxy for if it was installed using rv
+fn is_in_lock(pkg: &str, lock: Option<&Lockfile>) -> bool {
+    lock.map_or(true, |l| l.get_package(pkg, None).is_some())
+}
+
+fn is_binary_package(
     resolved_dep: &ResolvedDependency,
     repo_dbs: &[(RepositoryDatabase, bool)],
     r_version: &Version,
 ) -> bool {
-    if let Source::Repository { repository } = &resolved_dep.source {
-        if let Some((repo_db, _)) = repo_dbs.iter().find(|(db, _)| &db.url == repository) {
-            let version_requirement = Some(VersionRequirement::new(
-                resolved_dep.version.as_ref().clone(),
-                crate::package::Operator::Equal,
-            ));
-            repo_db
-                .find_package(
-                    &resolved_dep.name,
-                    version_requirement.as_ref(),
-                    r_version,
-                    resolved_dep.force_source,
-                )
-                .map(|(_, pkg_type)| pkg_type == PackageType::Binary)
-                .unwrap_or(false)
-        } else {
-            false
-        }
-    } else {
-        false
+    // We only will say a package is a binary if its from a repository
+    let repository = match &resolved_dep.source {
+        Source::Repository { repository } => repository,
+        _ => return false,
+    };
+    let ver_req = Some(VersionRequirement::new(
+        resolved_dep.version.as_ref().clone(),
+        Operator::Equal,
+    ));
+    repo_dbs
+        .iter()
+        .find(|(db, _)| &db.url == repository)
+        .and_then(|(db, _)| {
+            db.find_package(
+                &resolved_dep.name,
+                ver_req.as_ref(),
+                r_version,
+                resolved_dep.force_source,
+            )
+        })
+        .map(|(_, pkg)| pkg == PackageType::Binary)
+        .unwrap_or(false)
+}
+
+fn get_dep_id(dep: &ResolvedDependency, repos: &[Repository]) -> String {
+    match &dep.source {
+        Source::Repository { repository } => get_repository_alias(&repository, repos),
+        Source::Git { git, .. } => git.to_string(),
+        Source::Local { path } => path.to_string_lossy().to_string(),
+        Source::Url { url, .. } => url.to_string(),
     }
 }
 
-// Get an identifier for the source. For repositories, use the alias from the config. For git and url, use the url. For local, use the path
-fn get_source_id<'a>(resolved_dep: &'a ResolvedDependency, repos: &'a [Repository]) -> &'a str {
-    match &resolved_dep.source {
-        Source::Repository { repository } => repos
-            .iter()
-            .find(|r| r.url() == *repository)
-            .map_or(&repository, |r| &r.alias.as_str()),
-        Source::Local { path } => path.to_str().unwrap_or("local path"),
-        Source::Git { git, .. } => git,
-        Source::Url { url, .. } => url,
-    }
-}
-
-fn get_repo_alias<'a>(repo_db: &'a RepositoryDatabase, repos: &'a [Repository]) -> &'a str {
+fn get_repository_alias(r: &str, repos: &[Repository]) -> String {
     repos
         .iter()
-        .find(|r| r.url() == repo_db.url)
-        .map(|r| r.alias.as_str())
-        .unwrap_or("unknown")
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct RemoteInfo {
-    repositories: HashMap<String, (String, usize, usize)>,
-}
-
-impl RemoteInfo {
-    pub fn new(
-        r_version: &[u32; 2],
-        repository_databases: &[(RepositoryDatabase, bool)],
-        repos: &[Repository],
-    ) -> Self {
-        let mut repositories = HashMap::new();
-        let mut repo_id_hash = HashMap::new();
-        for (repo_db, force_source) in repository_databases {
-            let repo_id = repo_id_hash
-                .entry(&repo_db.url)
-                .or_insert(get_repo_alias(repo_db, repos));
-            let binary_count = if *force_source {
-                0
-            } else {
-                repo_db.get_binary_count(*r_version)
-            };
-            let source_count = repo_db.get_source_count();
-            repositories.insert(
-                repo_id.to_string(),
-                (repo_db.url.to_string(), binary_count, source_count),
-            );
-        }
-        Self { repositories }
-    }
-}
-
-impl fmt::Display for RemoteInfo {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let mut repos_vec: Vec<_> = self.repositories.iter().collect();
-        repos_vec.sort_by(|a, b| a.0.cmp(b.0));
-        for (alias, (url, binary_count, source_count)) in repos_vec {
-            write!(
-                f,
-                "  {alias} ({url}): {binary_count} binary packages, {source_count} source packages\n"
-            )?;
-        }
-        Ok(())
-    }
+        .find(|repo| repo.url() == r)
+        .map(|repo| repo.alias.to_string())
+        .unwrap_or(r.to_string()) 
 }
